Refactor imports in configuration files to use centralized paths
