+++ .github/repo-settings.dev.json.genie.ts
@@ -1,5 +1,5 @@
 import { requiredCIJobs } from '../genie/ci.ts'
-import { githubRuleset } from '../repos/effect-utils/genie/external.ts'
+import { githubRuleset } from '../genie/repo.ts'
 
 export default githubRuleset({
   name: 'dev-branch-rules',
