+++ .oxlintrc.json.genie.ts
@@ -1,9 +1,4 @@
-import {
-  baseOxlintCategories,
-  baseOxlintIgnorePatterns,
-  baseOxlintPlugins,
-} from './repos/effect-utils/genie/external.ts'
-import { oxlintConfig } from './repos/effect-utils/genie/external.ts'
+import { baseOxlintCategories, baseOxlintIgnorePatterns, baseOxlintPlugins, oxlintConfig } from './genie/repo.ts'
 
 /**
  * LiveStore oxlint configuration (Phase 1 — permissive).
