+++ megarepo.lock
@@ -4,9 +4,9 @@
     "effect-utils": {
       "url": "https://github.com/overengineeringstudio/effect-utils",
       "ref": "schickling/2026-03-08-pnpm-gvs-boundary-spec",
-      "commit": "a88e5d76d392e96e5b26c8d78d988afc8282cddc",
+      "commit": "6632ab1c548f74069017041b86a8f9de1419b888",
       "pinned": false,
-      "lockedAt": "2026-03-12T16:42:03.250Z"
+      "lockedAt": "2026-03-12T21:25:30.278Z"
     },
     "overeng-beads-public": {
       "url": "https://github.com/overengineeringstudio/overeng-beads-public",
